requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
defusedxml==0.6.0
Django==2.2.4
django-softdelete-it==0.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.8
idna==2.8
importlib-metadata==0.19
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
